feat(README): Enhancing the game section and adding Advanced Track link
The update brings a couple of changes, primarily to the README.md document.
Firstly, emojis were added to the titles of the analogy game and the "6️⃣ Degrees of Seperation" heading, providing visual cues for easier navigation through the markdown file :game_die:.
Also, few typographical errors were corrected; 'seperation' typo was corrected to 'separation' in the "6️⃣ Degrees of Separation" heading ensuring seamless readability.
Finally, an exciting new section was added as "Want More?" :rocket: which provides a link to the Advanced Track for users who want to delve deeper into the repository. This link provides users with a clear path to more complex content hosted from the same repository, facilitating community engagement.